% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_initial_mu_and_A.R
\name{calculate_initial_mu_and_A}
\alias{calculate_initial_mu_and_A}
\title{Calculate Initial Mu and A
 Helper function that, 
 given an expression matrix
 calculates mu_0 and A_0
 for the simulator:}
\usage{
calculate_initial_mu_and_A(E = matrix(ncol = 0, nrow = 0))
}
\arguments{
\item{E}{A custom expression matrix to use;  if not passed in, TCGA_PAAD data will be used}
}
\value{
A list containing mu_0 and A_0
}
\description{
\code{calculate_initial_mu_and_A} will take in a custom expression matrix, or use the TCGA_PAAD expression data to calculate initial mu and A
}
