% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalCOSDIST.R
\name{evalCOSDIST}
\alias{evalCOSDIST}
\title{Compare two sets of column data (automatic matching)}
\usage{
evalCOSDIST(truth, observed)
}
\description{
Compare two sets of column data (automatic matching)
}
