% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_W.R
\name{make_W}
\alias{make_W}
\title{Creates list of individual Ws for each sample and expression matrix from direchlet proportions in H and scRNA-seq data}
\usage{
make_W(data, H, total_cells)
}
\description{
Creates list of individual Ws for each sample and expression matrix from direchlet proportions in H and scRNA-seq data
}
