% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCentroid.R
\name{makeCentroid}
\alias{makeCentroid}
\title{Creates centroids representing average gene expression profile for each cell type
#'}
\usage{
makeCentroid(data)
}
\description{
Creates centroids representing average gene expression profile for each cell type
#'
}
