% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseGSE19830_LiverLungBrainRaw.R
\name{parseGSE19830_LiverLungBrainRaw}
\alias{parseGSE19830_LiverLungBrainRaw}
\title{get the liver lung brain data with getGEO}
\usage{
parseGSE19830_LiverLungBrainRaw()
}
\description{
get the liver lung brain data with getGEO
}
