% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_montrose.R
\name{pull_exprs_and_GEP}
\alias{pull_exprs_and_GEP}
\title{generate_data_set_list}
\usage{
pull_exprs_and_GEP(data = parsed_projects, gene = c("PHGDH", "PSAT1", "PSPH"))
}
\description{
generate_data_set_list
}
