% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange.R
\name{rearrange}
\alias{rearrange}
\title{Rearranges resultW, result H to match trueW, trueH ordering}
\usage{
rearrange(data, experiment)
}
\description{
Rearranges resultW, result H to match trueW, trueH ordering
}
