% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangeResult.R
\name{rearrangeResult}
\alias{rearrangeResult}
\title{Helper method to rearrange results to match truths}
\usage{
rearrangeResult(trueH, resultH, trueW, resultW)
}
\description{
Helper method to rearrange results to match truths
}
