% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Take in a single mu and A, and a mixture matrix H
Return mixed expression (after exponentiation)}
\usage{
sim_data(mu_tumor = 101 - (1:100), A_tumor = 1:100 \%*\% t(1:100),
  H = sim_H(), scaleFactor = rep(10, 3), seed = 1234)
}
\description{
Take in a single mu and A, and a mixture matrix H
Return mixed expression (after exponentiation)
}
