% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_1.R
\name{sim_data_1}
\alias{sim_data_1}
\title{Helper function that:
 use a single mixture and a set of mvn params 
 to simulate a single mixed sample}
\usage{
sim_data_1(h.i, mean, chol, seed = 1234)
}
\description{
Helper function that:
 use a single mixture and a set of mvn params 
 to simulate a single mixed sample
}
