% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_H.R
\name{sim_H}
\alias{sim_H}
\title{Take in direchelet parameters
Return a mixture matrix}
\usage{
sim_H(
  S = 10,
  seed = 1234,
  d.params = c(Tumor = 0.3, Stromal = 0.5, Immune = 0.1, Normal = 0.1)
)
}
\description{
Take in direchelet parameters
Return a mixture matrix
}
