% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_and_test.R
\name{simulate_and_test}
\alias{simulate_and_test}
\title{Description}
\usage{
simulate_and_test(A_tumor, mu_tumor, num.sim = 20, Samplesize = 100,
  scaleFactor = rep(10, 3), d.params = c(Tumor = 0.3, Stromal = 0.5,
  Immune = 0.1, Normal = 0.1), noise_setting = 2, seed = 1234)
}
\description{
Description
}
