% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetCentroid.R
\name{subsetCentroid}
\alias{subsetCentroid}
\title{Obtains cell-type specific gene expression profiles of X most common cell-types in data from centroid
#'}
\usage{
subsetCentroid(data, centroid, numTypes)
}
\description{
Obtains cell-type specific gene expression profiles of X most common cell-types in data from centroid
#'
}
